Change beer result to data class Added ViewModel Added RepoListItem
I commented parts that I can't solve
