+++ app/src/main/java/co/develhope/chooseyourownbeer/MainActivity.kt
@@ -13,6 +13,8 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import co.develhope.chooseyourownbeer.databinding.ActivityMainBinding
+import co.develhope.chooseyourownbeer.model.Beer
+import co.develhope.chooseyourownbeer.model.MainViewModel
 import co.develhope.chooseyourownbeer.network.BeersResult
 import co.develhope.chooseyourownbeer.network.PunkService
 import com.google.android.material.snackbar.Snackbar
@@ -22,22 +24,15 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 class MainActivity : AppCompatActivity() {
 
-    //Build retrofit
-    var retrofit = Retrofit.Builder()
-        .baseUrl("https://api.punkapi.com/v2/")
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    //Create class
-    var punkService = retrofit.create(PunkService::class.java)
-
     private lateinit var binding: ActivityMainBinding
+    private lateinit var viewModel: MainViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
 
         //Call method for beer repo
         retrievePunkRepos()
@@ -61,7 +56,6 @@ class MainActivity : AppCompatActivity() {
     fun retrievePunkRepos(){
         lifecycleScope.launch{
             try{
-                val repos = punkService.listRepos("beers")
                 //@TODO showRepos(repos)
             } catch (e: Exception){
                 Log.e("MainActivity", "Error retrieving repos: $e")
@@ -72,4 +66,12 @@ class MainActivity : AppCompatActivity() {
         }
     }
     //@TODO create function showRepos(repos)
+    fun showRepos(repoResults: BeersResult){
+        Log.d("MainActivity", "Repos received!")
+        val list = findViewById<RecyclerView>(R.id.repo_list_beer)
+        list.visibility = View.VISIBLE
+        //val adapter = BeerAdapter(repoResults)
+        //list.adapter = adapter
+    }
+
 }
