+++ app/src/main/java/co/develhope/chooseyourownbeer/network/BeersResult.kt
@@ -2,4 +2,6 @@ package co.develhope.chooseyourownbeer.network
 
 import co.develhope.chooseyourownbeer.model.Beer
 
-class BeersResult : ArrayList<Beer>()
+data class BeersResult (
+    var beersResultList: List<Beer>) {
+}
